main.go: improve the timeout calculation
Keep track for idle time on requests and input
separately then take the lower value of the two when comparing
with the target timeout.

This ensures that if for example the user was idle for a long
time but there were active requests, the timer will reset after
the requests have ended and not hibernate right away.